requirements.txt
acres==0.5.0
altair==5.5.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
astroid==3.2.4
attrs==25.4.0
bandit==1.8.6
black==25.9.0
blinker==1.9.0
build==1.3.0
cachetools==6.2.1
certifi==2025.10.5
cfgv==3.4.0
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.0
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
DAWG-Python==0.7.2
dill==0.4.0
distlib==0.4.0
Django==5.2.7
docopt==0.6.2
easyocr==1.7.2
etelemetry==0.3.1
faiss-cpu==1.12.0
fastapi==0.120.2
filelock==3.20.0
fitz==0.0.1.dev2
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.42.1
google-genai==1.47.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.36.0
identify==2.6.15
idna==3.11
imageio==2.37.0
iniconfig==2.3.0
isort==5.13.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keybert==0.9.0
lazy_loader==0.4
loguru==0.7.3
looseversion==1.3.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
mypy==1.18.2
mypy_extensions==1.1.0
narwhals==2.10.0
networkx==3.5
nibabel==5.3.2
ninja==1.13.0
nipype==1.10.0
nltk==3.9.2
nodeenv==1.9.1
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
pathlib==1.0.1
pathspec==0.12.1
pbr==7.0.1
pdf2image==1.17.0
pillow==12.0.0
pip-tools==7.5.1
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.3.0
protobuf==6.33.0
prov==2.1.1
puremagic==1.30
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
pydot==4.0.1
Pygments==2.19.2
pylint==3.2.7
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
PyMuPDF==1.26.5
pyparsing==3.2.5
pypdf==6.1.3
PyPDF2==3.0.1
pyproject_hooks==1.2.0
pytest==8.4.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytokens==0.2.0
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.3
rdflib==7.3.0
referencing==0.37.0
regex==2025.10.23
reportlab==4.4.4
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
ruff==0.14.2
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
shapely==2.1.2
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.49.1
stevedore==5.5.0
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
-e git+https://github.com/kelpi-e/Text2Sci.git@03b0c42ab42cb725f91df9cab03d6f7e99c7e7ca#egg=text2sci
threadpoolctl==3.6.0
tifffile==2025.10.16
tokenizers==0.22.1
toml==0.10.2
tomlkit==0.13.3
torch==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traits==7.0.2
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.35.4
watchdog==6.0.0
websockets==15.0.1
wheel==0.45.1
win32_setctime==1.2.0
